py-extract/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.build-system.sdist]
format = ["sdist", "wheel"]

[tool.setuptools]
include-package-data = true

[project]
name = "appleparser"
version = "1.0.0"
authors = [
    { name = "David Christie", email = "david.christiea@gmail.com" }
]
maintainers = [
    { name = "David Christie" }
]
description = "Lambda component that updates provider authentication cookies"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jmespath >=1, < 2",
    "requests >=2, < 3",
    "PyYAML >=6, < 7"
]